VDB-837 Update GetOrCreate functions to return id of type int64 (#210)
* Update id returned by GetOrCreate functions to be type int64
* Update all tests to expect int64
* Use `strconv.FormatInt` instead of Itoa so casting isn't necessary

+++ transformers/component_tests/queries/test_helpers/test_helpers.go
@@ -565,7 +565,7 @@ type FlipBid struct {
 	Tab   string
 }
 
-func SetUpFlipBidContext(setupData FlipBidContextInput) (ilkId, urnId int, err error) {
+func SetUpFlipBidContext(setupData FlipBidContextInput) (ilkId, urnId int64, err error) {
 	ilkId, ilkErr := shared.GetOrCreateIlk(setupData.IlkHex, setupData.Db)
 	if ilkErr != nil {
 		return 0, 0, ilkErr
